Add files via upload
For class Election, use the attributes defined in the class docstring, and ensure that the representation invariants hold at all times in between method calls.
For class Jurisdiction, use the attributes defined in the class docstring, and ensure that the representation invariants hold at all times in between method calls. 
Class Jurisdiction will use class Election heavily, but it must not access any of the attributes of class Election. They are all private.